+++ warden/x/warden/keeper/msg_server_new_sign_request.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	"context"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	types "github.com/warden-protocol/wardenprotocol/warden/x/warden/types/v1beta3"
@@ -29,25 +30,24 @@ func (k msgServer) NewSignRequest(ctx context.Context, msg *types.MsgNewSignRequ
 		return nil, err
 	}
 
-	if keychain.Fees != nil {
-		err := k.deductKeychainFees(
-			ctx,
-			msg.MaxKeychainFees,
-			keychain.AccAddress(),
-			keychain.Fees.SigReq,
-			sdk.MustAccAddressFromBech32(creator))
+	if err := keychain.EnsureSufficientSignFees(msg.MaxKeychainFees); err != nil {
+		return nil, err
+	}
 
-		if err != nil {
-			return nil, err
-		}
+	if err := k.deductKeychainFees(
+		ctx,
+		sdk.MustAccAddressFromBech32(creator),
+		keychain.Fees.SigReq); err != nil {
+		return nil, err
 	}
 
 	req := &types.SignRequest{
-		Creator:        creator,
-		KeyId:          msg.KeyId,
-		DataForSigning: msg.Input,
-		Status:         types.SignRequestStatus_SIGN_REQUEST_STATUS_PENDING,
-		EncryptionKey:  msg.EncryptionKey,
+		Creator:              creator,
+		KeyId:                msg.KeyId,
+		DataForSigning:       msg.Input,
+		Status:               types.SignRequestStatus_SIGN_REQUEST_STATUS_PENDING,
+		EncryptionKey:        msg.EncryptionKey,
+		DeductedKeychainFees: keychain.Fees.SigReq,
 	}
 
 	id, err := k.signRequests.Append(ctx, req)